Change way to display FeedTableViewCells
Now it should be the only one section with a multiple rows.

Add a special outlet for view inside the tableViewCell and change it's corner radius.

Also add a constraint between cell and cellView to get a little space between different articles.

Change the color of the selected cell to the clear color